+++ src/packageManager/ZipInstaller.ts
